+++ src/ZoneServer/Stomach.h
@@ -30,13 +30,13 @@ class Stomach
 		uint32	getFoodMax(){ return mFoodMax; }
 		void	setDrinkMax(uint32 drink){ mDrinkMax = drink; }
 		uint32	getDrinkMax(){ return mDrinkMax; }
-		void	setFood(double food){ mFood = food; gMessageLib->sendFoodUpdate(mParent);}
-		bool	checkFood(double food){ return !(mFood+ food >=mFoodMax); }
-		void	incFood(double food){ mFood += food; if(mFood < 0)mFood = 0; if(mFood > mFoodMax)mFood = mFoodMax; gMessageLib->sendFoodUpdate(mParent);}
+		void	setFood(float food){ mFood = food; gMessageLib->sendFoodUpdate(mParent);}
+		bool	checkFood(float food){ return !(mFood+ food >=mFoodMax); }
+		void	incFood(float food){ mFood += food; if(mFood < 0)mFood = 0.0f; if(mFood > mFoodMax)mFood = (float)mFoodMax; gMessageLib->sendFoodUpdate(mParent);}
 		uint32	getFood(){ return (uint32)mFood; }
-		void	setDrink(double drink){ mDrink = drink; gMessageLib->sendDrinkUpdate(mParent);}
-		bool	checkDrink(double drink){ return !(mDrink+ drink >=mDrinkMax); }
-		void	incDrink(double drink){ mDrink += drink; if(mDrink < 0)mDrink = 0; if(mDrink > mDrinkMax)mDrink = mDrinkMax; gMessageLib->sendDrinkUpdate(mParent);}
+		void	setDrink(float drink){ mDrink = drink; gMessageLib->sendDrinkUpdate(mParent);}
+		bool	checkDrink(float drink){ return !(mDrink+ drink >=mDrinkMax); }
+		void	incDrink(float drink){ mDrink += drink; if(mDrink < 0)mDrink = 0.0f; if(mDrink > mDrinkMax)mDrink = (float)mDrinkMax; gMessageLib->sendDrinkUpdate(mParent);}
 		uint32	getDrink(){ return (uint32)mDrink; }
 		bool	regenDrink(uint64 time,void*);
 		bool	regenFood(uint64 time,void*);
@@ -48,8 +48,8 @@ class Stomach
 		void checkForRegen();
 	private:
 		static const uint32  EmptyStomachTimer = 30; //Number of minutes before stomach clears
-		double	mFood;
-		double	mDrink;
+		float	mFood;
+		float	mDrink;
 		uint32	mFoodMax;
 		uint32	mDrinkMax;
 		PlayerObject* mParent;
